fix: fixing pytest errors due to updates in src/
